+++ klum-ast/src/test/groovy/com/blackbuild/groovy/configdsl/transform/TemplatesSpec.groovy
@@ -21,11 +21,15 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+//file:noinspection GrPackage
+//file:noinspection GroovyVariableNotAssigned
 package com.blackbuild.groovy.configdsl.transform
 
+import com.blackbuild.klum.ast.util.KlumInstanceProxy
 import org.codehaus.groovy.control.MultipleCompilationErrorsException
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
+import spock.lang.Ignore
 import spock.lang.Issue
 
 class TemplatesSpec extends AbstractDSLSpec {
@@ -185,8 +189,10 @@ class TemplatesSpec extends AbstractDSLSpec {
         }
 
         when:
+        def currentTemplates = null
         clazz.withTemplate(template) {
             instance = clazz.Create.With("Hallo") {
+                currentTemplates = KlumInstanceProxy.getProxyFor(delegate).currentTemplates
                 value "own"
             }
         }
@@ -195,6 +201,7 @@ class TemplatesSpec extends AbstractDSLSpec {
         instance.name == "Hallo"
         instance.value == "own"
         instance.value2 == "DefaultValue2"
+        currentTemplates == [(clazz): template]
     }
 
     def "Lists and Maps in template object should be cloned"() {
@@ -325,7 +332,7 @@ class TemplatesSpec extends AbstractDSLSpec {
 
         when:
         clazz.withTemplate(template) {
-            instance = getClass('pk.Child').Create.With()
+            instance = getClass('pk.Child').Create.One()
         }
 
         then:
@@ -799,6 +806,7 @@ class TemplatesSpec extends AbstractDSLSpec {
         instance.value == "DefaultValue"
     }
 
+    @Ignore
     def "locally applied templates using map"() {
         given:
         createClass('''
@@ -853,6 +861,7 @@ class TemplatesSpec extends AbstractDSLSpec {
         instance.value == null
     }
 
+    @Ignore
     def "locally applied templates with map"() {
         given:
         createClass('''
@@ -869,20 +878,20 @@ class TemplatesSpec extends AbstractDSLSpec {
                 String token
             }
         ''')
-        def fooClass = getClass("pk.Foo")
-        def fooTemplate = fooClass.Create.With(name: 'DefaultName')
-        def barClass = getClass("pk.Bar")
-        def barTemplate = barClass.Create.With(token: 'DefaultToken')
+        def Foo = getClass("pk.Foo")
+        def fooTemplate = Foo.Create.With(name: 'DefaultName')
+        def Bar = getClass("pk.Bar")
+        def barTemplate = Bar.Create.With(token: 'DefaultToken')
 
 
         def foo, bar
         when:
-        clazz.withTemplates((fooClass) : fooTemplate, (barClass) : barTemplate) {
-            foo = fooClass.Create.With {
+        clazz.withTemplates((Foo) : fooTemplate, (Bar) : barTemplate) {
+            foo = Foo.Create.With {
                 value 'blub'
             }
 
-            bar = barClass.Create.With()
+            bar = Bar.Create.One()
         }
 
         then:
@@ -919,7 +928,7 @@ class TemplatesSpec extends AbstractDSLSpec {
                 value 'blub'
             }
 
-            bar = barClass.Create.With()
+            bar = barClass.Create.One()
         }
 
         then:
@@ -958,6 +967,7 @@ class TemplatesSpec extends AbstractDSLSpec {
         bar.name == 'DefaultName'
     }
 
+    @Ignore
     def "parent child collections with map"() {
         given:
         createClass('''
@@ -1035,7 +1045,7 @@ class TemplatesSpec extends AbstractDSLSpec {
                 value 'blub'
             }
 
-            bar = barClass.Create.With()
+            bar = barClass.Create.One()
         }
 
         then:
@@ -1313,5 +1323,63 @@ class TemplatesSpec extends AbstractDSLSpec {
         template.key == null
         template.name == "Default"
         template.value == "DefaultValue"
+    }
+
+    @Issue("368")
+    def "instance proxies store the currently active templates"() {
+        given:
+        createClass('''
+            package pk
+
+import com.blackbuild.klum.ast.util.layer3.annotations.AutoCreate
+import com.blackbuild.klum.ast.util.KlumInstanceProxy
+
+            @DSL
+            class Foo {
+                String name
+                @Field(FieldType.TRANSIENT)
+                Map<Class, Object> activeTemplatesDuringAutoCreate
+                
+                @AutoCreate void storeTemplates() {
+                    activeTemplatesDuringAutoCreate = KlumInstanceProxy.getProxyFor(this).currentTemplates
+                }
+            }
+        ''')
+
+        when:
+        def template = clazz.Create.Template {
+            name "Default"
+        }
+
+        then:
+        template.name == "Default"
+
+        when:
+        def instance = clazz.Create.With {
+            name "Instance"
+        }
+        def proxy = KlumInstanceProxy.getProxyFor(instance)
+
+        then:
+        instance.name == "Instance"
+        proxy.currentTemplates == [:]
+
+        when:
+        def templatesDuringCreation = null
+        clazz.withTemplate(template) {
+            instance = clazz.Create.With {
+                name "Overridden"
+                templatesDuringCreation = KlumInstanceProxy.getProxyFor(delegate).currentTemplates
+            }
+        }
+        proxy = KlumInstanceProxy.getProxyFor(instance)
+
+        then:
+        instance.name == "Overridden"
+        templatesDuringCreation == [(clazz): template]
+        instance.activeTemplatesDuringAutoCreate == [(clazz): template]
+
+        and: "templates are cleared after the instance is created"
+        proxy.currentTemplates[clazz] == null
     }
 }